--- toyml/classification/naive_bayes/__init__.py
@@ -1,9 +1,0 @@
-from .categorical_naive_bayes import CategoricalNaiveBayes
-from .gaussian_naive_bayes import GaussianNaiveBayes
-from .multinomial_naive_bayes import MultinomialNaiveBayes
-
-__all__ = [
-    "GaussianNaiveBayes",
-    "MultinomialNaiveBayes",
-    "CategoricalNaiveBayes",
-]
